Reduce the scope of SCHED_LOCK for a per-proc lock
Introduce per-proc mutex to protect proc's private state, instead of
using SCHED_LOCK for that.

Lock order:

  process mutex --> proc mutex --> SCHED_LOCK

That works on the way to sleep. Wakeup and the scheduler come from the
other direction. That results in dealing with reverse lock order. In
addition, the release order may be different from the acquire order.
Internally these mutexes maintain an SPL, where release order matters.
Put an explicit spl*(9) around these cases where it matters to maintain
proper priority levels.

The current proc and the proc chosen to be next by the scheduler are
both locked during cpu_switchto(9). Therefore cpu_switchto(9) now
returns the previous proc to unlock it. If the previous proc exit(3)ed
it is not accessible by the scheduler anymore and it unlocks itself in
the exit path. cpu_switchto(9) returns NULL in that case.

The new design calls mi_switch() with the proc lock held, instead of
SCHED_LOCK. Access to sleep queues remains SCHED_LOCK'ed.

The sysctl(2)s on proc internal state remain unlocked for now.