btrace: Shell quirk for command line forwarding
Allow an executable btrace script to forward its command line parameters
to btrace(8) itself. That is done with a shebang of /bin/sh and exec'ing
btrace in the second line. So the bt(5) script starts with

 |#!/bin/sh
 |exec /usr/sbin/btrace "$@" "$0"
 |
 |...

That means btrace needs to ignore the sh(1) exec command on the second
line.