btrace(8) & ELF entry points
Hello Christian,

Below is a diff that cleans a bit Tom's work to add dynamic entry points
to every entry in the ELF symbol table of the kernel for use with btrace(8).

The code is currently enabled with the DDBPROF option because it has
been developed and tested as an alternative for a kernel compiled with
GPROF option.

Would you be interested to continue polishing this diff?

Cheers,
Martin