exit: Wakeup parent early
Wakeup the parent process early in exit1(). This is safe now that the
scheduler makes sure to notify the parent after the child is off of the
CPU. The parent now waits for that notification before it frees the
child's resources.

That means the reaper does not handle zombie processes any longer.