exit: Unlock part of process_zap()
The lower part of the function is freeing resources back to the
allocators. These operations are MP-safe. Split off the upper part so
the KERNEL_UNLOCK/KERNEL_LOCK dance is better visible.

Note that proc_free() is run without kernel lock in the reaper. That can
run in parallel already.