Synchromize the reaper from the scheduler
The previous approach to get to a safe place for the reaper to free a
dead proc's resources was a local CPU list that idle has put on the
global list of dead procs for the reaper.

Change that synchronization to a conditional signal that the scheduler
sends to the reaper immediately after the dead proc is off the CPU.

This decouples waking up the reaper (or the parent in the future) from
waiting for the final schedule event.