exit: Set proc in SDEAD state later
When the proc dies, we have to make sure that none of its reaources are
freed asynchronously while it still executes. We have this covered with
a dead condition now. That meand the proc is eligible to sleep for
longer in the exit path.

The process is not allowed to sleep anmore at cpu_exit(). Conceptually,
the architecture code may leave the CPU in a state where switching
contexts is not possible any longer.